# sdk/python/contract/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RecordEventRequest": {
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "event_type": {
            "minLength": 1,
            "title": "Event Type",
            "type": "string"
          },
          "graph_state_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph State Hash"
          },
          "payload": {
            "title": "Payload",
            "type": "object"
          },
          "workflow_id": {
            "minLength": 1,
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "event_type"
        ],
        "title": "RecordEventRequest",
        "type": "object"
      },
      "RecordEventResponse": {
        "properties": {
          "event_id": {
            "title": "Event Id",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "event_id",
          "workflow_id"
        ],
        "title": "RecordEventResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WorkflowEventsResponse": {
        "properties": {
          "events": {
            "items": {
              "type": "object"
            },
            "title": "Events",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "events",
          "total"
        ],
        "title": "WorkflowEventsResponse",
        "type": "object"
      },
      "WorkflowListResponse": {
        "properties": {
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowSummary"
            },
            "title": "Workflows",
            "type": "array"
          }
        },
        "required": [
          "workflows",
          "total"
        ],
        "title": "WorkflowListResponse",
        "type": "object"
      },
      "WorkflowSummary": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "event_count": {
            "title": "Event Count",
            "type": "integer"
          },
          "first_event": {
            "title": "First Event",
            "type": "string"
          },
          "last_event": {
            "title": "Last Event",
            "type": "string"
          },
          "workflow_id": {
            "title": "Workflow Id",
            "type": "string"
          }
        },
        "required": [
          "workflow_id",
          "agent_id",
          "first_event",
          "last_event",
          "event_count"
        ],
        "title": "WorkflowSummary",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "SpAIder \u2014 A multi-agent semantic knowledge graph platform powered by LLMs and Neo4j. Ingest unstructured text, build a persistent knowledge graph, query it in natural language, and synthesise fine-tuning datasets for downstream models.",
    "title": "SpAIder",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/replay/events": {
      "post": {
        "description": "Record a workflow event for later replay/audit.",
        "operationId": "record_event_api_v1_replay_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Event",
        "tags": [
          "replay"
        ]
      }
    },
    "/api/v1/replay/workflows": {
      "get": {
        "description": "List distinct workflow runs available for replay.",
        "operationId": "list_workflows_api_v1_replay_workflows_get",
        "parameters": [
          {
            "description": "Filter by agent ID",
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by agent ID",
              "title": "Agent Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Workflows",
        "tags": [
          "replay"
        ]
      }
    },
    "/api/v1/replay/workflows/{workflow_id}/events": {
      "get": {
        "description": "Fetch the complete event log for a specific workflow run.",
        "operationId": "get_workflow_events_api_v1_replay_workflows__workflow_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "workflow_id",
            "required": true,
            "schema": {
              "title": "Workflow Id",
              "type": "string"
            }
          },
          {
            "description": "Filter events after this ISO timestamp",
            "in": "query",
            "name": "start_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter events after this ISO timestamp",
              "title": "Start Time"
            }
          },
          {
            "description": "Filter events before this ISO timestamp",
            "in": "query",
            "name": "end_time",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter events before this ISO timestamp",
              "title": "End Time"
            }
          },
          {
            "description": "Comma-separated event type filter",
            "in": "query",
            "name": "event_types",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated event type filter",
              "title": "Event Types"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 1000,
              "maximum": 10000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowEventsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow Events",
        "tags": [
          "replay"
        ]
      }
    },
...
  }
}